+++ src/components/main/main.tsx
@@ -1,5 +1,6 @@
 import React, { PropsWithChildren } from 'react'
 import Header from '../layout/header'
+import { Breadcrumb } from '../layout'
 interface MainProps extends PropsWithChildren {
   title?: React.ReactNode
   icon?: React.ReactNode
@@ -14,6 +15,9 @@ export const Main = ({ title, icon, header, actions, children }: MainProps) => {
       </Header>
       <div className='flex-1 overflow-auto'>
         <div className='p-4'>
+          <div className='mb-4'>
+            <Breadcrumb />
+          </div>
           <div>{children}</div>
         </div>
       </div>